Split flex into two file for panel/layout
